requirements.txt
langchain
langchain-community
langchain-core
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
openai==1.59.8
python-dotenv==1.0.1
fastapi==0.111.1
langserve==0.2.2
sse_starlette==2.1.2
langchain_chroma
pytest==8.2.2
langchain-google-genai~=2.1.8
tavily-python==0.3.4
langgraph==0.1.8
beautifulsoup4==4.12.3
langchain==0.2.7
langchainhub==0.1.20
langchain-community==0.2.7
langchain-chroma==0.1.2
langchain-openai==0.1.16
